refactor(chat): extract message-preparation pipeline into prepare-model-messages
