+++ platform/backend/src/routes/chat/routes.test.ts
@@ -26,6 +26,7 @@ import ConversationModel from "@/models/conversation";
 import MessageModel from "@/models/message";
 import { test } from "@/test";
 import type { ChatMessage } from "@/types";
+import { __test as __prepareTest } from "./prepare-model-messages";
 import {
   __test,
   buildChatStopConditions,
@@ -37,7 +38,7 @@ import {
 
 describe("prepareMessagesForProvider", () => {
   it("normalizes csv files to text/plain for anthropic", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "anthropic",
       messages: [
         {
@@ -63,7 +64,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("normalizes markdown files to text/plain for anthropic", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "anthropic",
       messages: [
         {
@@ -89,7 +90,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("normalizes json files to text/plain for anthropic", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "anthropic",
       messages: [
         {
@@ -127,7 +128,7 @@ describe("prepareMessagesForProvider", () => {
       ],
     };
 
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "openai",
       messages: [message],
     });
@@ -143,7 +144,7 @@ describe("prepareMessagesForProvider", () => {
   };
 
   it("prepends placeholder text for bedrock user messages with only a file part", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [{ role: "user", parts: [pdfFilePart] }],
     });
@@ -155,7 +156,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("prepends placeholder when the only existing text part is whitespace", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [
         {
@@ -177,7 +178,7 @@ describe("prepareMessagesForProvider", () => {
       parts: [{ type: "text", text: "Summarize this" }, pdfFilePart],
     };
 
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [message],
     });
@@ -186,7 +187,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("pads bedrock assistant messages whose only text part is whitespace", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [{ role: "assistant", parts: [{ type: "text", text: "" }] }],
     });
@@ -200,7 +201,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("pads bedrock messages whose reasoning lacks a bedrock signature", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [
         {
@@ -219,7 +220,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("pads bedrock messages that only contain ignored UI data parts", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [
         {
@@ -247,7 +248,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("pads bedrock messages that only contain step markers and ignored data parts", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [
         {
@@ -272,7 +273,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("pads bedrock messages that only contain streaming tool input", () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [
         {
@@ -299,7 +300,7 @@ describe("prepareMessagesForProvider", () => {
   });
 
   it("pads empty bedrock assistant step blocks before later tool calls", async () => {
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [
         {
@@ -357,7 +358,7 @@ describe("prepareMessagesForProvider", () => {
       ],
     };
 
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [message],
     });
@@ -377,7 +378,7 @@ describe("prepareMessagesForProvider", () => {
       ],
     };
 
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [message],
     });
@@ -397,7 +398,7 @@ describe("prepareMessagesForProvider", () => {
       ],
     };
 
-    const messages = __test.prepareMessagesForProvider({
+    const messages = __prepareTest.prepareMessagesForProvider({
       provider: "bedrock",
       messages: [message],
     });
@@ -411,7 +412,7 @@ describe("buildModelMessagesForProvider", () => {
   const conversationId = "conv-model-prep";
 
   it("drops an assistant turn that converts to empty model content", async () => {
-    const modelMessages = await __test.buildModelMessagesForProvider({
+    const modelMessages = await __prepareTest.buildModelMessagesForProvider({
       provider: "openai",
       conversationId,
       messages: [
@@ -435,7 +436,7 @@ describe("buildModelMessagesForProvider", () => {
   });
 
   it("keeps normal text and tool assistant turns", async () => {
-    const modelMessages = await __test.buildModelMessagesForProvider({
+    const modelMessages = await __prepareTest.buildModelMessagesForProvider({
       provider: "openai",
       conversationId,
       messages: [
